requirements.txt
# Core Dependencies (Python 3.8 compatible)
numpy>=1.24.0,<1.25.0
pandas>=2.0.0,<2.1.0
scikit-learn>=1.3.0,<1.4.0
scipy>=1.10.0,<1.11.0

# Deep Learning
torch>=1.13.0,<2.1.0
torchvision>=0.14.0,<0.16.0
timm>=0.6.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Explainability
shap>=0.41.0

# Data Download
kaggle>=1.5.12

# Utilities
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.28.0

# Optional: For enhanced visualizations
plotly>=5.9.0
